김초욱 공격로 버프 미적용 버그 수정
공격로 버프 미적용 버그 수정

+++ PlumTowerDefence/Assets/Scripts/Enemy/Enemy.cs
@@ -500,7 +500,7 @@ public class Enemy : MonoBehaviour
         Debug.Log(laneBufftype);
 
         currentBuffType = laneBufftype;
-        float statChange =Tables.MonsterLaneBuff.Get((int)currentBuffType)._Amount;
+        float statChange =Tables.MonsterLaneBuff.Get(currentBuffType)._Amount;
 
         switch (currentBuffType)
         {
